ll->int

+++ kyopro/library/datastructure/unionfind.cpp
@@ -5,10 +5,10 @@
 
 //0-indexed
 struct unionfind {
-	vector<ll> par, siz;
+	vector<int> par, siz;
 
-	unionfind(ll n) : par(n), siz(n) {
-		for (ll i = 0; i < n; ++i) {
+	unionfind(int n) : par(n), siz(n) {
+		for (int i = 0; i < n; ++i) {
 			//全部根で初期化
 			par[i] = i;
 			//サイズは1
@@ -16,10 +16,10 @@ struct unionfind {
 		}
 	}
 
-	void init(ll n) {
+	void init(int n) {
 		par.resize(n);
 		siz.resize(n);
-		for (ll i = 0; i < n; ++i) {
+		for (int i = 0; i < n; ++i) {
 			//全部根で初期化
 			par[i] = i;
 			//サイズは1
@@ -28,13 +28,13 @@ struct unionfind {
 	}
 
 	//根を返す
-	ll find(ll a) {
+	int find(int a) {
 		return par[a] == a ? a : par[a] = find(par[a]);
 	}
 
 	//くっつける。元から同じだったらfalseを返す
-	bool unite(ll a, ll b) {
-		ll x = find(a), y = find(b);
+	bool unite(int a, int b) {
+		int x = find(a), y = find(b);
 		if (x == y)return false;
 		else if (siz[x] < siz[y]) {
 			par[x] = y;
@@ -52,20 +52,20 @@ struct unionfind {
 	}
 
 	//同じ集合か判定する
-	bool same(ll a, ll b) {
+	bool same(int a, int b) {
 		return find(a) == find(b);
 	}
 
 	//サイズを返す
-	ll size(ll a) {
+	int size(int a) {
 		return siz[find(a)];
 	}
 
 	//同じ集合に属す葉を纏めて返す
-	vector<ll> leaf(ll a) {
-		vector<ll> x;
-		ll n = par.size();
-		for (ll i = 0; i < n; ++i)if (same(a, i))x.push_back(i);
+	vector<int> leaf(int a) {
+		vector<int> x;
+		int n = par.size();
+		for (int i = 0; i < n; ++i)if (same(a, i))x.push_back(i);
 		return x;
 	}
